Move Coordinates to CoordinatesBattle
