+++ app/src/Template/CoordinatesBattle/battle.ctp
