+++ app/src/Template/CoordinatesBattle/favorite.ctp
