+++ app/src/Template/CoordinatesBattle/get_new_coordinate.ctp
