+++ app/src/Template/CoordinatesBattle/get_score.ctp
