+++ app/src/Template/CoordinatesBattle/result.ctp
