+++ app/src/Template/CoordinatesBattle/select_battle_mode.ctp
